Implemented StationsWidget
