+++ e_metrobus/navigation/constants.py
@@ -1,0 +1,9 @@
+
+
+STATIONS = [
+    "Hertzallee",
+    "S+U Zoologischer Garten",
+    "S+U Zoologischer Garten/Jebensstr.",
+    "Breitscheidplatz",
+    "Budapester Str."
+]
