+++ e_metrobus/templates/navigation/route.html
@@ -3,121 +3,11 @@
 
 {% block content %}
 
-<section class="grid-container" id="stationsList">
-
-	<div class="grid-x station" id="station01">
-		<div class="cell small-2 station__line">
-			<svg width="52" height="50" id="stationLine1">
-				<!-- COMPONENT START -->
-				<rect x="20" y="0" width="12" height="50" />
-				<circle x="20" y="0" cx="26" cy="25" r="4" fill="#ffffff" />
-				<!-- COMPONENT END -->
-			</svg>
-			<!-- COMPONENT START -->
-			<svg width="52" height="50" style="display: none">
-				<polygon points="22,19 34,25 22,31" style="fill:#000000;" />
-			</svg>
-			<!-- COMPONENT END -->
-		</div>
-		<div class="cell small-10 station__name">
-			Hertzallee
-		</div>
-	</div>
-
-	<div class="grid-x">
-		<div class="cell station__between" id="betweenStations1">
-			<!-- COMPONENT START -->
-			<svg width="52" height="10">
-				<rect x="20" y="0" width="12" height="10" />
-			</svg>
-			<!-- COMPONENT END -->
-		</div>
-	</div>
-
-	<div class="grid-x station" id="station02">
-		<div class="cell small-2 station__line">
-			<svg width="52" height="50" id="stationLine2">
-				<rect x="20" y="0" width="12" height="50" />
-				<circle x="20" y="0" cx="26" cy="25" r="4" fill="#ffffff" />
-			</svg>
-			<svg width="52" height="50" style="display:none">
-				<polygon points="22,19 34,25 22,31" style="fill:#000000;" />
-			</svg>
-		</div>
-		<div class="cell small-10 station__name">
-			S+U Zoologischer Garten
-		</div>
-	</div>
-
-	<div class="grid-x">
-		<div class="cell station__between" id="betweenStations2">
-			<svg width="52" height="10">
-				<rect x="20" y="0" width="12" height="10" />
-			</svg>
-		</div>
-	</div>
-
-	<div class="grid-x station" id="station03">
-		<div class="cell small-2 station__line">
-			<svg width="52" height="50" id="stationLine3">
-				<rect x="20" y="0" width="12" height="50" />
-				<circle x="20" y="0" cx="26" cy="25" r="4" fill="#ffffff" />
-			</svg>
-			<svg width="52" height="50" style="display:none">
-				<polygon points="22,19 34,25 22,31" style="fill:#000000;" />
-			</svg>
-		</div>
-		<div class="cell small-10 station__name">
-			S+U Zoologischer Garten/Jebensstr.
-		</div>
-	</div>
-
-	<div class="grid-x">
-		<div class="cell station__between" id="betweenStations3">
-			<svg width="52" height="10">
-				<rect x="20" y="0" width="12" height="10" />
-			</svg>
-		</div>
-	</div>
-
-	<div class="grid-x station" id="station04">
-		<div class="cell small-2 station__line">
-			<svg width="52" height="50" id="stationLine4">
-				<rect x="20" y="0" width="12" height="50" />
-				<circle x="20" y="0" cx="26" cy="25" r="4" fill="#ffffff" />
-			</svg>
-			<svg width="52" height="50" style="display:none">
-				<polygon points="22,19 34,25 22,31" style="fill:#000000;" />
-			</svg>
-		</div>
-		<div class="cell small-10 station__name">
-			Breitscheidplatz
-		</div>
-	</div>
-
-	<div class="grid-x">
-		<div class="cell station__between" id="betweenStations4">
-			<svg width="52" height="10">
-				<rect x="20" y="0" width="12" height="10" />
-			</svg>
-		</div>
-	</div>
-
-	<div class="grid-x station" id="station05">
-		<div class="cell small-2 station__line">
-			<svg width="52" height="50" id="stationLine5">
-				<rect x="20" y="0" width="12" height="50" />
-				<circle x="20" y="0" cx="26" cy="25" r="4" fill="#ffffff" />
-			</svg>
-			<svg width="52" height="50" style="display:none">
-				<polygon points="22,19 34,25 22,31" style="fill:#000000;" />
-			</svg>
-		</div>
-		<div class="cell small-10 station__name">
-			Budapester Str.
-		</div>
-	</div>
-
-</section>
+{{stations}}
 
 {% endblock %}
+
+{% block javascript %}
+  {{block.super}}
+  {{stations.media}}
+{% endblock %}
