+++ e_metrobus/templates/widgets/stations.html
@@ -1,0 +1,36 @@
+<section class="grid-container" id="stationsList">
+
+  {% for station in stations %}
+    <div class="grid-x station" id='station{{forloop.counter|stringformat:"02d"}}'>
+      <div class="cell small-2 station__line">
+        <svg width="52" height="50" id="stationLine{{forloop.counter}}">
+          <!-- COMPONENT START -->
+          <rect x="20" y="0" width="12" height="50" />
+          <circle x="20" y="0" cx="26" cy="25" r="4" fill="#ffffff" />
+          <!-- COMPONENT END -->
+        </svg>
+        <!-- COMPONENT START -->
+        <svg width="52" height="50" style="display: none">
+          <polygon points="22,19 34,25 22,31" style="fill:#000000;" />
+        </svg>
+        <!-- COMPONENT END -->
+      </div>
+      <div class="cell small-10 station__name">
+        {{station}}
+      </div>
+    </div>
+
+    {% if not forloop.last %}
+      <div class="grid-x">
+        <div class="cell station__between" id="betweenStations{{forloop.counter}}">
+          <!-- COMPONENT START -->
+          <svg width="52" height="10">
+            <rect x="20" y="0" width="12" height="10" />
+          </svg>
+          <!-- COMPONENT END -->
+        </div>
+      </div>
+    {% endif %}
+  {% endfor %}
+
+</section>
